+++ lib/authentication/auth_gate.dart
@@ -4,15 +4,27 @@ import '../home_screen.dart';
 import 'register_or_login.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 
-class AuthenticationGate extends StatelessWidget {
+class AuthenticationGate extends StatefulWidget {
   const AuthenticationGate({super.key});
+
+  @override
+  State<AuthenticationGate> createState() => _AuthenticationGateState();
+}
+
+class _AuthenticationGateState extends State<AuthenticationGate> {
+  // Cached so Firebase re-emitting auth state on browser focus events doesn't remount children
+  late final Stream<User?> _authStream;
+
+  @override
+  void initState() {
+    super.initState();
+    _authStream = authService.value.authStateChanges;
+  }
 
   @override
   Widget build(BuildContext context) {
     return StreamBuilder<User?>(
-      stream: authService
-          .value
-          .authStateChanges, // Check if there is a logged-in user
+      stream: _authStream, // Check if there is a logged-in user
       builder: (context, snapshot) {
         if (snapshot.connectionState == ConnectionState.waiting) {
           // waiting for the check